backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.4
boilerpy3==1.0.7
boto3==1.35.76
botocore==1.35.76
cattrs==24.1.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.6
configobj==5.0.6

constantly==15.1.0
cryptography==3.4.8
Django==5.0.6
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.1
filelock==3.14.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflect==7.4.0
iniconfig==2.0.0
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
keyring==23.5.0
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.55.0
packaging==24.0
pandas==2.2.3
pillow==11.0.0
posthog==3.7.2
psycopg2-binary==2.9.10
PyJWT==2.3.0
pyOpenSSL==21.0.0
pytest==8.2.1
python-dotenv==1.0.1
pytz==2022.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
SQLAlchemy==2.0.36
sqlparse==0.5.0
sympy==1.13.1
tenacity==9.0.0
tqdm==4.67.0
transformers==4.46.3
Twisted==22.1.0
typing_extensions==4.12.1
urllib3==1.26.5
virtualenv==20.26.2
wikipedia==1.4.0
yarl==1.17.2

